requirements.txt
anyio==4.0.0
attrs==22.2.0
certifi==2023.7.22
dnspython==2.6.1
exceptiongroup==1.1.0
flake8==6.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pymongo==4.6.1
pytest==7.2.2
python-dotenv==1.0.0
python-telegram-bot==20.5
sniffio==1.3.0
tablib==3.3.0
tomli==2.0.1
xlrd==2.0.1
xlwt==1.3.0
